requirements.txt
pydantic==1.8.2
websockets==9.0.2
pika==1.2.0
websocket_client==1.0.1

redis==3.5.3
asyncio==3.4.3
APScheduler==3.7.0